renamed file and component name
