+++ client/src/components/BoardDetails.jsx
